requirements.txt
Flask>=2.0
rembg
Pillow
